+++ NEWS.md
@@ -22,6 +22,7 @@ New Features
   - `"rhombille"` implements a rhombille tiling of rhombi (#37)
   - `"tetrakis_square"` implements a tetrakis square tiling of isosceles right triangles (#38)
   - `"2*.2**.2*.2**"` implements a polygon tiling made of rhombi.
+  - `"2**.3**.12*"` implements a polygon tiling made of rhombi, triangles, and twelve-pointed stars.
   - `"3.3*.3.3*"` implements a regular (star) polygon tiling made of triangles and three-pointed stars.
   - `"3.3.3.12*.3.3.12*"` implements a regular (star) polygon tiling made of triangles and twelve-pointed starts.
   - `"3.3.8*.3.4.3.8*"` implements a regular (star) polygon tiling of triangles,